# packages/contracts/openapi/school-os-api.yaml
openapi: 3.1.0
info:
  title: School OS API
  version: "1.0"
  description: >
    Versioned public/partner-facing surface of the School OS platform.
    Authoritative implementation lives in apps/platform (Laravel);
    this file is the source of truth for its shape and is consumed by
    packages/shared-types to generate TypeScript bindings. See
    docs/architecture/API.md for the full architecture (auth, pagination,
    idempotency, rate limiting, webhooks, error format).
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /schools/{schoolId}/enrollments:
    get:
      operationId: listEnrollments
      summary: >
        Administrative Enrollment directory across the School. Requires
        enrollments.view.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { name: academic_year_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: campus_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: grade_level_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: section_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [active, completed, withdrawn, transferred, cancelled] } }
        - { name: student_number, in: query, required: false, schema: { type: string } }
        - { name: student_name, in: query, required: false, schema: { type: string } }
        - { name: roll_number, in: query, required: false, schema: { type: string } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/EnrollmentSummary" } }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
  /schools/{schoolId}/enrollments/{enrollmentId}:
    get:
      operationId: getEnrollment
      summary: Fetch one Enrollment's detail. Requires enrollments.view.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/EnrollmentId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Enrollment" } } }
        "404":
          description: Not found (including a cross-School id -- indistinguishable from nonexistent).
...
  /schools/{schoolId}/students/{studentId}/enrollments:
    get:
      operationId: listStudentEnrollmentHistory
      summary: Full historical Enrollment record for one Student. Requires enrollments.view.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/StudentId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { type: array, items: { $ref: "#/components/schemas/Enrollment" } } } }
        "404":
          description: Not found (including a cross-School Student id -- indistinguishable from nonexistent).
    post:
      operationId: createStudentEnrollment
      summary: >
        Enrolls a Student into a Section. AcademicYear/Campus/GradeLevel
        are always derived from the given Section -- never accepted
        independently. Requires enrollments.manage. Idempotency-Key
        required.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/StudentId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollmentInput" }
      responses:
        "201":
          description: Created.
          headers:
            Idempotency-Replayed: { $ref: "#/components/headers/IdempotencyReplayed" }
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/Enrollment" } } }
        "422":
          description: >
            A second active Enrollment for this Student+Year, a
            duplicate roll number in this Section+Year, or a blank
            roll number.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /schools/{schoolId}/enrollment-rollovers:
    get:
      operationId: listEnrollmentRollovers
      summary: >
        Administrative rollover Plan directory. Requires
        enrollments.view AND enrollments.rollovers.view (both,
        independent -- CapabilityResolver has no inheritance).
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { name: source_academic_year_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: target_academic_year_id, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [draft, validated, executing, completed, completed_with_errors, cancelled] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/RolloverPlanSummary" } }
                  meta: { $ref: "#/components/schemas/PaginationMeta" }
    post:
      operationId: createEnrollmentRollover
      summary: >
        Creates a draft rollover Plan. Requires enrollments.manage AND
        enrollments.rollovers.manage. Idempotency-Key required.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RolloverPlanInput" }
      responses:
        "201":
          description: Created
          headers:
            Idempotency-Replayed: { $ref: "#/components/headers/IdempotencyReplayed" }
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/RolloverPlanSummary" } } }
        "422":
          description: >
            Validation failure (including a foreign-School or
            nonexistent Academic Year id -- indistinguishable), same
            source/target year, or an open Plan already exists for this
            year pair.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /schools/{schoolId}/enrollment-rollovers/{rollover}:
    get:
      operationId: getEnrollmentRollover
      summary: >
        Rollover Plan detail -- metadata, Mappings, and a durable
        execution summary. Requires enrollments.view AND
        enrollments.rollovers.view. Never embeds the (potentially very
        large) Item collection; see the separate paginated Items
        endpoint.
      parameters:
        - { $ref: "#/components/parameters/SchoolId" }
        - { $ref: "#/components/parameters/RolloverId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, required: [data], properties: { data: { $ref: "#/components/schemas/RolloverPlan" } } }
        "404":
          description: Not found (including a cross-School id -- indistinguishable from nonexistent).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: >
        Opaque personal-access token (web/mobile) or partner API key
        (third-party integrations). See docs/architecture/API.md
        ("Authentication").
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Required on all unsafe (POST/PATCH/DELETE) mutating requests that
        are not naturally idempotent, e.g. payment callbacks. See
        docs/architecture/API.md ("Idempotency").
      schema: { type: string }
    SchoolId:
      name: schoolId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    StudentId:
      name: studentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EnrollmentId:
      name: enrollmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    RolloverId:
      name: rollover
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
...
    IdempotencyReplayed:
      description: >
        Present and "true" only when this response is a replay of a
        previously completed request with the same Idempotency-Key
        (docs/architecture/API.md, "Idempotency"); absent on a request's
        first, freshly-executed response.
      schema: { type: string, enum: ["true"] }
...
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, perPage, total]
      properties:
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, status, requestId]
          properties:
            message: { type: string }
            status: { type: integer }
            code:
              type: string
              nullable: true
              description: >
                Stable machine-readable error code (e.g.
                IDEMPOTENCY_KEY_CONFLICT), present only for exceptions
                that define one. null otherwise -- see
                docs/architecture/API.md ("Error format").
            requestId: { type: string, format: uuid, nullable: true }
            errors:
              type: object
              nullable: true
              description: Field-level validation errors, present only on 422 responses.
...
    StudentSummary:
      type: object
      required: [id, studentNumber, firstName, status]
      description: >
        List-row shape -- deliberately excludes dateOfBirth (Highly
        Sensitive children's data, docs/security/DATA-CLASSIFICATION.md:
        minimized default visibility in broad list views). See Student
        for the full single-record shape.
      properties:
        id: { type: string, format: uuid }
        studentNumber: { type: string }
        firstName: { type: string }
        middleName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        status: { type: string, enum: [active, inactive] }
...
    EnrollmentRef:
      type: object
      required: [id, name, code]
      description: A minimal AcademicYear/Campus/GradeLevel/Section reference embedded in an Enrollment row.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
    EnrollmentSummary:
      type: object
      required: [id, student, academicYear, campus, gradeLevel, section, rollNumber, status, startsOn]
      description: >
        Shared shape for directory rows AND single-record detail (there
        is no broader "full" shape) -- deliberately excludes Student
        dateOfBirth and all Guardian data
        (docs/security/DATA-CLASSIFICATION.md, "minimized default
        visibility"). rollNumber is always a string, never coerced to
        a number (e.g. "007" stays "007").
      properties:
        id: { type: string, format: uuid }
        student: { $ref: "#/components/schemas/StudentSummary" }
        academicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        campus: { $ref: "#/components/schemas/EnrollmentRef" }
        gradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        section: { $ref: "#/components/schemas/EnrollmentRef" }
        rollNumber: { type: string }
        status: { type: string, enum: [active, completed, withdrawn, transferred, cancelled] }
        startsOn: { type: string, format: date }
        endsOn: { type: string, format: date, nullable: true }
    Enrollment:
      allOf:
        - { $ref: "#/components/schemas/EnrollmentSummary" }
        - type: object
          required: [createdAt, updatedAt]
          properties:
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    EnrollmentInput:
      type: object
      required: [section_id, roll_number, starts_on]
      description: >
        The Section is the sole authoritative placement input --
        academic_year_id/campus_id/grade_level_id/school_id/student_id
        are never accepted here and are always derived server-side from
        section_id (StudentEnrollmentService::enroll()).
      properties:
        section_id: { type: string, format: uuid }
        roll_number: { type: string, maxLength: 255 }
        starts_on: { type: string, format: date }

    # Phase 8A.14 -- HR read-only transport schemas. Deliberately
    # snake_case property names, UNLIKE every schema above this line --
    # this accurately reflects the REAL, already-committed JSON shape
    # every App\Domain\HR\Application\*::toArray() DTO has produced
    # since 8A.8/8A.9/8A.10/8A.11 (this file is the source of truth for
    # the SHAPE the implementation already has, not a license to
    # rename it to match an unrelated module's casing convention).
...
    RolloverPlanSummary:
      type: object
      required: [id, sourceAcademicYear, targetAcademicYear, status, configurationVersion, createdAt]
      description: >
        Phase 1B.7E. Directory-row shape -- deliberately omits mappings
        and executionSummary (only present on RolloverPlan detail) to
        keep the directory listing cheap regardless of Plan count.
      properties:
        id: { type: string, format: uuid }
        sourceAcademicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        targetAcademicYear: { $ref: "#/components/schemas/EnrollmentRef" }
        status: { type: string, enum: [draft, validated, executing, completed, completed_with_errors, cancelled] }
        configurationVersion: { type: integer }
        validatedConfigurationVersion: { type: integer, nullable: true }
        createdBy: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        createdAt: { type: string, format: date-time }
        validatedAt: { type: string, format: date-time, nullable: true }
        executionStartedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
    RolloverMapping:
      type: object
      required: [id, sourceGradeLevel, targetGradeLevel, isRepeat]
      properties:
        id: { type: string, format: uuid }
        sourceGradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        sourceSection: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        targetGradeLevel: { $ref: "#/components/schemas/EnrollmentRef" }
        targetSection: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        isRepeat: { type: boolean }
    RolloverSubjectOfferingRef:
      type: object
      required: [id, status]
      description: >
        Phase 1G.4. A minimal, PII-free SubjectOffering reference for
        the subject-mapping workspace -- Subject name/code and academic
        dimensions (GradeLevel/Campus/ElectiveGroup) are non-PII School
        reference/configuration data. `status` is always present so a
        durable mapping's target can be displayed (with an inactive
        warning) even after the Offering itself later became inactive
        (docs/students/PHASE-1G-4-SUBJECT-ROLLOVER-API-UI.md).
      properties:
        id: { type: string, format: uuid }
        subject: { allOf: [{ $ref: "#/components/schemas/EnrollmentRef" }], nullable: true }
        gradeLevel: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        campus: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
        status: { type: string, enum: [active, inactive] }
        electiveGroup: { type: object, nullable: true, properties: { id: { type: string, format: uuid }, name: { type: string } } }
    RolloverSubjectMapping:
      type: object
      required: [id, sourceSubjectOffering, state]
      description: >
        Phase 1G.4/1G.1. `state` is the three-state model surfaced for
        display: this schema only ever represents a row that EXISTS
        (`mapped` or `omit`) -- the UNCONFIGURED state is the ABSENCE
        of a row, represented separately by
        `RolloverPlan.unmappedSourceSubjectOfferings`, never collapsed
        into this list.
      properties:
        id: { type: string, format: uuid }
        sourceSubjectOffering: { $ref: "#/components/schemas/RolloverSubjectOfferingRef" }
        state: { type: string, enum: [mapped, omit] }
        targetSubjectOffering: { allOf: [{ $ref: "#/components/schemas/RolloverSubjectOfferingRef" }], nullable: true }
...
    RolloverPlan:
      allOf:
        - { $ref: "#/components/schemas/RolloverPlanSummary" }
        - type: object
          required: [isValidatedForCurrentConfiguration, mappings, subjectMappings, unmappedSourceSubjectOfferings, executionSummary]
          properties:
            isValidatedForCurrentConfiguration: { type: boolean }
            mappings: { type: array, items: { $ref: "#/components/schemas/RolloverMapping" } }
            subjectMappings: { type: array, items: { $ref: "#/components/schemas/RolloverSubjectMapping" } }
            unmappedSourceSubjectOfferings:
              type: array
              description: >
                Phase 1G.4. Bounded discovery list of elective source
                SubjectOfferings with current eligible participation in
                this Plan but NO mapping row yet -- what would trigger
                `missing_subject_mapping` in dry-run. Never implies a
                mapping row was created; purely a read-only projection
                (EnrollmentRolloverReadService::unmappedSourceSubjectOfferings()).
              items: { $ref: "#/components/schemas/RolloverSubjectOfferingRef" }
            executionSummary: { $ref: "#/components/schemas/RolloverExecutionSummary" }
    RolloverPlanInput:
      type: object
      required: [source_academic_year_id, target_academic_year_id]
      description: >
        school_id/status/configuration_version/validated_configuration_version/created_by_user_id/execution
        timestamps are never accepted -- School derives from the route,
        actor from the authenticated User, every other field is
        EnrollmentRolloverPlanService::createDraft()'s own invariant.
      properties:
        source_academic_year_id: { type: string, format: uuid }
        target_academic_year_id: { type: string, format: uuid }
...
    RolloverExecutionSummary:
      type: object
      required: [total, succeeded, reconciled, skipped, failed, pending]
      description: >
        Recomputed from persisted Items every time -- never an
        in-memory counter. `planStatus` is present only on start()/
        resume()'s own response (EnrollmentRolloverExecutionService's
        return value), not on the Plan-detail-embedded executionSummary.
      properties:
        total: { type: integer }
        succeeded: { type: integer }
        reconciled: { type: integer }
        skipped: { type: integer }
        failed: { type: integer }
        pending: { type: integer }
        planStatus: { type: string, enum: [draft, validated, executing, completed, completed_with_errors, cancelled] }
...